extension/requirements.txt
requests
pyperclip
Pillow